Revert "emit uint256 cluster balance"
This reverts commit 4c4a9dc3d58c3b50bdbe3f40bf837d4c0da70072.

+++ contracts/ISSVNetworkCore.sol
@@ -39,7 +39,7 @@ interface ISSVNetworkCore {
         uint64 networkFee;
         uint64 networkFeeIndex;
         uint64 index;
-        uint256 balance;
+        uint64 balance;
         bool disabled;
     }
